project update

+++ controllers/clothingItems.js
@@ -1,9 +1,16 @@
 const ClothingItem = require("../models/clothingItem");
-const { BAD_REQUEST, NOT_FOUND, SERVER_ERROR } = require("../utils/errors");
+const {
+  OK,
+  CREATED,
+  BAD_REQUEST,
+  FORBIDDEN,
+  NOT_FOUND,
+  SERVER_ERROR,
+} = require("../utils/errors");
 
-const getItem = (req, res) => {
+const getClothingItem = (req, res) => {
   ClothingItem.find({})
-    .then((item) => res.status(200).send(item))
+    .then((item) => res.status(OK).send(item))
     .catch((error) => {
       console.error(error);
       return res
@@ -12,11 +19,11 @@ const getItem = (req, res) => {
     });
 };
 
-const createItem = (req, res) => {
+const createClothingItem = (req, res) => {
   const { name, weather, imageUrl } = req.body;
 
   ClothingItem.create({ name, weather, imageUrl, owner: req.user._id })
-    .then((item) => res.status(201).send(item))
+    .then((item) => res.status(CREATED).send(item))
     .catch((error) => {
       console.error(error);
       if (error.name === "ValidationError") {
@@ -28,19 +35,31 @@ const createItem = (req, res) => {
     });
 };
 
-const deleteItem = (req, res) => {
+const deleteClothingItem = (req, res) => {
   const { itemId } = req.params;
+  const itemOwner = req.user._id;
 
-  ClothingItem.findByIdAndDelete(itemId)
-    .orFail()
-    .then((item) => res.status(200).send(item))
+  ClothingItem.findById(itemId)
+    .then((item) => {
+      if (!item) {
+        throw new Error("NotFound");
+      }
+      if (item.owner.toString() !== itemOwner) {
+        throw new Error("Forbidden");
+      }
+      return ClothingItem.findByIdAndDelete(itemId);
+    })
+    .then((item) => res.status(OK).send(item))
     .catch((error) => {
       console.error(error);
+      if (error.message === "NotFound") {
+        return res.status(NOT_FOUND).send({ message: error.message });
+      }
+      if (error.message === "Forbidden") {
+        return res.status(FORBIDDEN).send({ message: error.message });
+      }
       if (error.name === "CastError") {
         return res.status(BAD_REQUEST).send({ message: error.message });
-      }
-      if (error.name === "DocumentNotFoundError") {
-        return res.status(NOT_FOUND).send({ message: error.message });
       }
       return res
         .status(SERVER_ERROR)
@@ -55,7 +74,7 @@ const likeItem = (req, res) => {
     { new: true }
   )
     .orFail()
-    .then((item) => res.status(200).send(item))
+    .then((item) => res.status(OK).send(item))
     .catch((error) => {
       console.error(error);
       if (error.name === "CastError") {
@@ -77,7 +96,7 @@ const dislikeItem = (req, res) => {
     { new: true }
   )
     .orFail()
-    .then((item) => res.status(200).send(item))
+    .then((item) => res.status(OK).send(item))
     .catch((error) => {
       console.error(error);
 
@@ -94,9 +113,9 @@ const dislikeItem = (req, res) => {
 };
 
 module.exports = {
-  createItem,
-  getItem,
-  deleteItem,
+  getClothingItem,
+  createClothingItem,
+  deleteClothingItem,
   likeItem,
   dislikeItem,
 };
